+++ footer.php
@@ -18,13 +18,12 @@
         </div><!-- .footer-bar -->
         <div class="row">
                 <div class="small-6 large-6 columns">
-                    <p class="copyright">&copy; 2016 | Copyright 2016 Samuel Guebo.<br><a href="<?php echo esc_url( __( 'https://wordpress.org/', 'acajou' ) ); ?>"><?php printf( esc_html__( 'Proudly powered by %s', 'acajou' ), 'WordPress' ); ?></a> and available on <a href="https://github.com/samuelguebo/acajou"><i class="fa fa-github"></i> Github</a></p>
+                    <p class="copyright"><?php _e('Developed by ','acajou')?>Samuel Guebo.<br><a href="<?php echo esc_url( __( 'https://wordpress.org/', 'acajou' ) ); ?>"><?php printf( esc_html__( 'Proudly powered by %s', 'acajou' ), 'WordPress' ); ?></a> <?php _e('and available on','acajou')?> <a href="https://github.com/samuelguebo/acajou"><i class="fa fa-github"></i> Github</a></p>
                 </div>
 
                 <div class="small-2 large-2 columns socials right">
                     <ul>
-                        <li><a href="#"><i class="fa fa-twitter"></i></a></li>
-                        <li><a href="#"><i class="fa fa-facebook"></i></a></li>
+                        <?php acajou_social_links();?>
                     </ul>
                 </div>
         </div><!-- copyright-->