+++ inc/customizer.php
@@ -11,9 +11,85 @@
  * @param WP_Customize_Manager $wp_customize Theme Customizer object.
  */
 function acajou_customize_register( $wp_customize ) {
-	$wp_customize->get_setting( 'blogname' )->transport         = 'postMessage';
-	$wp_customize->get_setting( 'blogdescription' )->transport  = 'postMessage';
-	$wp_customize->get_setting( 'header_textcolor' )->transport = 'postMessage';
+	
+    // Create sections for socials links
+    $wp_customize->add_section("social_links", array(
+		"title" => __("Social links", "acajou"),
+		"priority" => 30,
+	));
+    
+    // Facebook link
+    $wp_customize->add_setting("facebook_url", array(
+		"default" => "",
+		"transport" => "refresh",
+	));
+    $wp_customize->add_control(new WP_Customize_Control(
+		$wp_customize,
+		"facebook_url",
+		array(
+			"label" => __("Enter Facebook url", "acajou"),
+			"section" => "social_links",
+			"settings" => "facebook_url",
+			"type" => "text",
+		)
+	));
+    
+    // Twitter link
+    $wp_customize->add_setting("twitter_url", array(
+		"default" => "",
+		"transport" => "refresh",
+	));
+    $wp_customize->add_control(new WP_Customize_Control(
+		$wp_customize,
+		"twitter_url",
+		array(
+			"label" => __("Enter Twitter url", "acajou"),
+			"section" => "social_links",
+			"settings" => "twitter_url",
+			"type" => "text",
+		)
+	));
+    
+    // Youtube link
+    $wp_customize->add_setting("youtube_url", array(
+		"default" => "",
+		"transport" => "refresh",
+	));
+    $wp_customize->add_control(new WP_Customize_Control(
+		$wp_customize,
+		"youtube_url",
+		array(
+			"label" => __("Enter Youtube url", "acajou"),
+			"section" => "social_links",
+			"settings" => "youtube_url",
+			"type" => "text",
+		)
+	));
+    
+    // Github link
+    $wp_customize->add_setting("github_url", array(
+		"default" => "",
+		"transport" => "refresh",
+	));
+    $wp_customize->add_control(new WP_Customize_Control(
+		$wp_customize,
+		"github_url",
+		array(
+			"label" => __("Enter Github url", "acajou"),
+			"section" => "social_links",
+			"settings" => "github_url",
+			"type" => "text",
+		)
+	));
+    
+    
+    /*
+     * Theme colors
+     *
+     */
+    
+
+    
 }
 add_action( 'customize_register', 'acajou_customize_register' );
 
@@ -23,4 +99,3 @@ add_action( 'customize_register', 'acajou_customize_register' );
 function acajou_customize_preview_js() {
 	wp_enqueue_script( 'acajou_customizer', get_template_directory_uri() . '/js/customizer.js', array( 'customize-preview' ), '20151215', true );
 }
-add_action( 'customize_preview_init', 'acajou_customize_preview_js' );