+++ template-parts/content-article.php
@@ -23,7 +23,7 @@
         <div class="panel">
             <h6 class="post-item-title"><a href="<?php the_permalink();?>"><?php the_title();?></a></h6>
             <p><?php the_excerpt();?></p> 
-            <a href="<?php the_permalink();?>" class="small button post-item-buttom">Lire la suite</a> 
+            <a href="<?php the_permalink();?>" class="small button post-item-buttom"><?php _e('Read more','acajou')?></a> 
             <?php if ( get_edit_post_link() ) : ?>
                 <a href="<?php echo get_edit_post_link();?>" class="small button alert"><i class="fa fa-edit"></i> edit</a>