+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.0.0] - 2016-12-30
 ### Added
-- Custom background image
+- Custom header image(s)
 - Set social links in Customizer
 - Six flavors: chocolate, blue, green, red, purple, and yellow. You might switch between these colors in using the Customizer.
 - Add Fading effects to block using [Scroll Reveal](https://github.com/jlmakes/scrollreveal) library